requirements.txt
# Local STT Dependencies
faster-whisper
pyaudio
numpy
python-dotenv

# LiveKit Demo Dependencies (from the document)
livekit-agents
livekit-plugins-assemblyai
livekit
python-dotenv 